can pause and resume the training, better status feedback

+++ packages/server/src/training/index.ts
@@ -14,6 +14,12 @@ const TRAIN_IMAGE = "gcperkins/wpilib-ml-train:latest";
 const TEST_IMAGE = "gcperkins/wpilib-ml-test:latest";
 const CONTAINER_MOUNT_PATH = "/opt/ml/model";
 
+//training state enumeration
+const NOT_TRAINING = 0;
+const PREPARING = 1;
+const TRAINING = 2;
+const PAUSED = 3;
+
 type TrainParameters = {
   "eval-frequency": number;
   "dataset-path": string[];
@@ -119,7 +125,7 @@ export default class Trainer {
           test: null
         },
         status: {
-          trainingInProgress: false,
+          trainingState: NOT_TRAINING,
           currentEpoch: 0,
           lastEpoch: EPOCHS
         }
@@ -165,7 +171,7 @@ export default class Trainer {
           test: null
         },
         status: {
-          trainingInProgress: false,
+          trainingState: NOT_TRAINING,
           currentEpoch: 0,
           lastEpoch: project.epochs
         }
@@ -177,9 +183,12 @@ export default class Trainer {
     const datasets = await project.getDatasets();
     if (!datasets) {
       Promise.reject("there are no datasets. How am I supposed to train with no datasets?");
+      return;
     }
 
     const ID = project.id;
+    this.projects[ID].status.trainingState = PREPARING;
+
     const MOUNT = this.projects[ID].directory;
     const DATASETPATHS = datasets.map((dataset) =>
       path.posix.join(CONTAINER_MOUNT_PATH, "dataset", path.basename(dataset.path))
@@ -200,29 +209,30 @@ export default class Trainer {
     };
     await fs.promises.writeFile(path.posix.join(MOUNT, "hyperparameters.json"), JSON.stringify(trainParameters));
 
-
+    if (!this.projects[ID].status.trainingState) return "training stopped";
     this.projects[ID].containers.metrics = await this.createContainer(METRICS_IMAGE, "METRICS-", ID, MOUNT, "6006");
-    const OLD_TRAIN_DIR = path.posix.join(MOUNT,"train")
-    if (fs.existsSync(OLD_TRAIN_DIR)){
+    const OLD_TRAIN_DIR = path.posix.join(MOUNT, "train");
+    if (fs.existsSync(OLD_TRAIN_DIR)) {
       fs.rmdirSync(OLD_TRAIN_DIR, { recursive: true });
+      console.log("old train dir removed");
     } //if this project has already trained, we must get rid of the evaluation files in order to only get new metrics
-    const OLD_METRICS_DIR = path.posix.join(MOUNT,"metrics.json")
+    const OLD_METRICS_DIR = path.posix.join(MOUNT, "metrics.json");
     if (fs.existsSync(OLD_METRICS_DIR)) {
       fs.unlinkSync(OLD_METRICS_DIR);
     } //must clear old checkpoints in order for new ones to be saved by trainer
     this.projects[ID].checkpoints = {}; //must add a way to preserve existing checkpoints somehow
-    
-    this.projects[ID].status.currentEpoch = 0;
-    this.projects[ID].status.trainingInProgress = true;
 
+    if (!this.projects[ID].status.trainingState) return "training stopped";
     datasets.forEach((dataset) => {
       fs.copyFileSync(
         path.posix.join("data", dataset.path),
         path.posix.join(MOUNT, "dataset", path.basename(dataset.path))
       );
     });
+    console.log("datasets copied");
 
     //custom checkpoints not yet supported by gui
+    if (!this.projects[ID].status.trainingState) return "training stopped";
     if (project.initialCheckpoint != "default") {
       if (!fs.existsSync(path.posix.join(MOUNT, "checkpoints"))) {
         await mkdirp(path.posix.join(MOUNT, "checkpoints"));
@@ -243,20 +253,24 @@ export default class Trainer {
       ]);
     }
 
-    if (!this.projects[ID].status.trainingInProgress) return "training stopped";
+    if (!this.projects[ID].status.trainingState) return "training stopped";
+    console.log("extracting the dataset");
     this.projects[ID].containers.train = await this.createContainer(DATASET_IMAGE, "TRAIN-", ID, MOUNT);
     await this.projects[ID].containers.train.start();
     await this.projects[ID].containers.train.wait();
     await this.projects[ID].containers.train.remove();
+    console.log("datasets extracted");
 
-    if (!this.projects[ID].status.trainingInProgress) return "training stopped";
+    if (!this.projects[ID].status.trainingState) return "training stopped";
+    this.projects[ID].status.trainingState = TRAINING;
+    this.projects[ID].status.currentEpoch = 0;
     this.projects[ID].containers.train = await this.createContainer(TRAIN_IMAGE, "TRAIN-", ID, MOUNT);
     await this.projects[ID].containers.metrics.start();
     await this.projects[ID].containers.train.start();
     await this.projects[ID].containers.train.wait();
     await this.projects[ID].containers.train.remove();
 
-    this.projects[ID].status.trainingInProgress = false;
+    this.projects[ID].status.trainingState = NOT_TRAINING;
     this.projects[ID].containers.train = null;
     return "training complete";
   }
@@ -269,6 +283,7 @@ export default class Trainer {
 
     if (!fs.existsSync(path.posix.join(MOUNT, "train", `${CHECKPOINT_TAG}.meta`))) {
       Promise.reject("cannot find requested checkpoint");
+      return;
     }
 
     await mkdirp(path.posix.join(EXPORT_PATH, "checkpoint"));
@@ -334,9 +349,11 @@ export default class Trainer {
 
     if (!fs.existsSync(modelExport.tarPath)) {
       Promise.reject("model not found");
+      return;
     }
     if (!fs.existsSync(videoPath)) {
       Promise.reject("video not found");
+      return;
     }
     if (!fs.existsSync(MOUNTED_MODEL_PATH)) {
       await fs.promises.copyFile(modelExport.tarPath, MOUNTED_MODEL_PATH);
@@ -394,11 +411,41 @@ export default class Trainer {
   }
 
   async halt(id: string): Promise<void> {
-    this.projects[id].status.trainingInProgress = false;
     if (this.projects[id].containers.train) {
       if ((await this.projects[id].containers.train.inspect()).State.Running) {
         await this.projects[id].containers.train.kill({ force: true });
       }
+    }
+    this.projects[id].status.trainingState = NOT_TRAINING;
+    Promise.resolve();
+  }
+
+  async toggleContainer(id: string, pause: boolean): Promise<void> {
+    if (!this.projects[id].status.trainingState) {
+      console.log("not training");
+      Promise.resolve();
+      return;
+    }
+    if (!this.projects[id].containers.train) {
+      console.log("no container to pause");
+      Promise.resolve();
+      return;
+    }
+    const CONTAINER = this.projects[id].containers.train;
+
+    switch (pause) {
+      case true:
+        if (!(await CONTAINER.inspect()).State.Paused) {
+          await CONTAINER.pause();
+          this.projects[id].status.trainingState = PAUSED;
+        }
+        break;
+      case false:
+        if ((await CONTAINER.inspect()).State.Paused) {
+          CONTAINER.unpause();
+          this.projects[id].status.trainingState = TRAINING;
+        }
+        break;
     }
     Promise.resolve();
   }
